% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data60UK.R
\name{catalogueData60UK}
\alias{catalogueData60UK}
\title{Data source: Data60UK catalogue}
\usage{
catalogueData60UK(bbox = NULL, columnName = NULL, columnValue = NULL,
  cached = TRUE)
}
\arguments{
\item{bbox}{bounding box, a list made of 4 elements: minimum longitude (lonMin), minimum latitude (latMin), maximum longitude (lonMax), maximum latitude (latMax)}

\item{columnName}{name of the column to filter}

\item{columnValue}{value to look for in the column named columnName}

\item{cached}{boolean value. Default is TRUE (use cached datasets).}
}
\value{
This function returns a data frame made of 61 rows (gauging stations) and 6 columns: "id" (hydrometric reference number), "River", "Location", "gridReference", "Latitude", "Longitude".
}
\description{
This function interfaces the Data60UK database catalogue (available from http://www.nwl.ac.uk/ih/nrfa/pub/index.html) containing 61 datasets. Dataset catalogue is available from \url{http://www.nwl.ac.uk/ih/nrfa/pub/data.html}.
}
\examples{
\dontrun{
  # Retrieve the whole catalogue
  x <- catalogueData60UK()

  # Define a bounding box
  bbox <- list(lonMin = -4, latMin = 52, lonMax = -2, latMax = 53)

  # Filter the catalogue
  x <- catalogueData60UK(bbox)
  x <- catalogueData60UK(columnName = "id", columnValue = "62001")
}

}
\author{
Claudia Vitolo
}

