% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOPEX.R
\name{catalogueMOPEX}
\alias{catalogueMOPEX}
\title{Data source: MOPEX catalogue}
\usage{
catalogueMOPEX()
}
\value{
This function returns a data frame made of 431 rows (gauging stations) and 12 columns containing stations metadata.
}
\description{
This function interfaces the MOPEX database catalogue (available from \url{ftp://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/}) containing 438 daily datasets.
}
\examples{
\dontrun{
  # Retrieve the MOPEX catalogue
  x <- catalogueMOPEX()
}

}
\author{
Claudia Vitolo
}

