% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data60UKCatalogue}
\alias{data60UKCatalogue}
\title{Interface for the Data60UK database catalogue}
\usage{
data60UKCatalogue(BBlonMin = -180, BBlonMax = +180, BBlatMin = -90,
  BBlatMax = +90)
}
\arguments{
\item{BBlonMin}{Minimum latitude of bounding box}

\item{BBlonMax}{Maximum latitude of bounding box}

\item{BBlatMin}{Minimum longitude of bounding box}

\item{BBlatMax}{Maximum longitude of bounding box}
}
\value{
This function returns a data frame made of 5 columns: "id" (hydrometric reference number), "name", "location", "Latitude" and "Longitude".
}
\description{
This function interfaces the Data60UK database catalogue (available from http://www.nwl.ac.uk/ih/nrfa/pub/index.html) containing 61 datasets. Dataset catalogue is available from http://www.nwl.ac.uk/ih/nrfa/pub/data.html.
}
\examples{
# data60UKCatalogue()
# data60UKCatalogue(BBlonMin=-3.82,BBlonMax=-3.63,BBlatMin=52.41,BBlatMax=52.52)
}
\author{
Claudia Vitolo
}

