% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContent}
\alias{getContent}
\title{Extracts links from ftp page}
\usage{
getContent(dirs)
}
\arguments{
\item{dirs}{is the url from which links should be extracted}
}
\value{
vector containing all the links in the page
}
\description{
This function extracts all the links in a ftp page
}
\examples{
# Retrieve mopex daily catalogue
# url <- "ftp://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/Us_438_Daily"
# getContent(url)

}
\author{
Claudia Vitolo
}

