% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GRDCCatalogue}
\alias{GRDCCatalogue}
\title{Interface for the Global Runoff Data Centre database catalogue}
\usage{
GRDCCatalogue(lonMin = -180, lonMax = +180, latMin = -90, latMax = +90,
  liveData = FALSE, mdDescription = FALSE)
}
\arguments{
\item{lonMin}{Minimum latitude of bounding box}

\item{lonMax}{Maximum latitude of bounding box}

\item{latMin}{Minimum longitude of bounding box}

\item{latMax}{Maximum longitude of bounding box}

\item{liveData}{bolean value that allows to switch between the online dataset (liveData=TRUE) and the cached one (liveData = FALSE). The online database retrieval might be slower but more up-to-date compared with the cached one (updated twice a year).}

\item{mdDescription}{boolean value. Default is FALSE (no description is printed)}
}
\value{
list of stations within the bounding box
}
\description{
This function interfaces the Global Runoff Data Centre database which provides river discharge data for about 9000 sites over 157 countries.
}
\examples{
# GRDCCatalogue(lonMin=-3.82,lonMax=-3.63,latMin=52.43,latMax=52.52)
}
\author{
Claudia Vitolo
}

