% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtools-package.R
\docType{data}
\name{grdcLTMMD}
\alias{grdcLTMMD}
\title{Data set: The grdcLTMMD look-up table}
\format{
A data frame with 6 rows and 4 columns.
\describe{
  \item{\code{WMO_Region}}{an integer between 1 and 6}
  \item{\code{Coverage}}{}
  \item{\code{Number_of_stations}}{}
  \item{\code{Archive}}{url to spreadsheet}
}
}
\source{
\url{http://www.bafg.de/GRDC}
}
\usage{
data("grdcLTMMD")
}
\description{
The grdcLTMMD look-up table
}
\keyword{datasets}
