% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadDAILY.R
\name{HadDAILY}
\alias{HadDAILY}
\title{Interface for the HadUKP - UK regional precipitation series}
\usage{
HadDAILY()
}
\value{
list of 11 time series (1 per region/subregion)
}
\description{
This function interfaces the Met Office Hadley Centre observations datasets: HadUKP which provides daily precipitation data for 3 regions and 8 subregions in the United Kingdom.
}
\examples{
\dontrun{
  # Retrieve the full dataset (list of 11 time series)
  HadDAILY_all <- HadDAILY()
  plot(x$EWP) # time series for England and Wales
}

}
\author{
Claudia Vitolo
}

