% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtools-package.R
\docType{package}
\name{hddtools}
\alias{hddtools}
\alias{hddtools-package}
\title{hddtools: Hydrological Data Discovery Tools}
\description{
: Many governmental bodies and institutions are currently committed to publish open data as the result of a trend of increasing transparency, based on which a wide variety of information produced at public expense is now becoming open and freely available to improve public involvement in the process of decision and policy making. Discovery, access and retrieval of information is, however, not always a simple task. Especially when access to data APIs is not allowed, downloading metadata catalogue, select the information needed, request datasets, de-compression, conversion, manual filtering and parsing can become rather tedious. The R package hddtools is an open source project, designed to make all the above operations more efficient by means of reusable functions. The package facilitate access to various online data sources such as the Global Runoff Data Centre, NASA's TRMM mission, the Data60UK database amongst others. This package complements R's growing functionality in environmental web technologies to bridge the gap between data providers and data consumers and it is designed to be the starting building block of scientific workflows for linking data and models in a seamless fashion.
}
\details{
Facilitates discovery and handling of hydrological data, access to catalogues and databases.
}
\references{
Vitolo C, Buytaert W, 2014, HDDTOOLS: an R package serving Hydrological Data
Discovery Tools, AGU Fall Meeting, 15-19 December 2014, San Francisco, USA.
}

