% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mopex.R
\name{mopexCatalogue}
\alias{mopexCatalogue}
\title{Interface for the MOPEX database catalogue}
\usage{
mopexCatalogue(bbox = NULL, metadataColumn = NULL, entryValue = NULL,
  verbose = FALSE)
}
\arguments{
\item{bbox}{bounding box, a list made of 4 elements: minimum longitude (lonMin), minimum latitude (latMin), maximum longitude (lonMax), maximum latitude (latMax)}

\item{metadataColumn}{name of the column to filter}

\item{entryValue}{value to look for in the column named metadataColumn}

\item{verbose}{if TRUE it returns whether the data is coming from live or cached data sources}
}
\value{
This function returns a data frame made of 5 columns: "id" (hydrometric reference number), "name", "location", "Latitude" and "Longitude".
}
\description{
This function interfaces the MOPEX database catalogue (available from ftp://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/) containing 438 daily datasets.
}
\examples{
# Retrieve the whole catalogue
# mopexCatalogue()
}
\author{
Claudia Vitolo
}

