% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mopex.R
\name{mopexTS}
\alias{mopexTS}
\title{Interface for the MOPEX database of Daily Time Series}
\usage{
mopexTS(hydroRefNumber, plotOption = FALSE, timeExtent = NULL)
}
\arguments{
\item{hydroRefNumber}{hydrometric reference number}

\item{plotOption}{boolean to define whether to plot the results. By default this is set to TRUE.}

\item{timeExtent}{is a vector of dates and times for which the data should be retrieved}
}
\value{
The function returns a data frame containing 2 time series (as zoo objects): "P" (precipitation) and "Q" (discharge).
}
\description{
This function extract the dataset containing daily rainfall and streamflow discharge at one of the MOPEX locations.
}
\examples{
# mopexTS("14359000")
}
\author{
Claudia Vitolo
}

