% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trmm}
\alias{trmm}
\title{Download and convert TRMM data}
\usage{
trmm(fileLocation, url)
}
\arguments{
\item{fileLocation}{file ptah where to save the GeoTiff}

\item{url}{url where data is stored (e.g. "ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/3B43_V7/2012/")}
}
\value{
Data is loaded as rasterbrick, then converted to a multilayer Geotiff that can
}
\description{
The TRMM dataset provide global hystorical rainfall estimation in a gridded format.
}
\details{
This code is based upon the following blog post:
TRMM FAQ: http://disc.sci.gsfc.nasa.gov/additional/faq/precipitation_faq.shtml
}
\examples{
# trmm(fileLocation,url)
}
\author{
Claudia Vitolo
}

