% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data60UK.R
\name{tsData60UK}
\alias{tsData60UK}
\title{Interface for the Data60UK database of Daily Time Series}
\usage{
tsData60UK(hydroRefNumber, plotOption = FALSE, twindow = NULL)
}
\arguments{
\item{hydroRefNumber}{hydrometric reference number (string)}

\item{plotOption}{boolean to define whether to plot the results. By default this is set to TRUE.}

\item{twindow}{is a vector of dates and times for which the data should be retrieved}
}
\value{
The function returns a data frame containing 2 time series (as zoo objects): "P" (precipitation) and "Q" (discharge).
}
\description{
This function extract the dataset containing daily rainfall and streamflow discharge at one of the Data60UK locations.
}
\examples{
\dontrun{
  tsData60UK(hydroRefNumber = 39015)
}

}
\author{
Claudia Vitolo
}

