% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRDC.R
\name{catalogueGRDC}
\alias{catalogueGRDC}
\title{Data source: Global Runoff Data Centre catalogue}
\usage{
catalogueGRDC(bbox = NULL, stationID = NULL, columnName = NULL,
  columnValue = NULL, mdDescription = FALSE, cached = TRUE)
}
\arguments{
\item{bbox}{bounding box, a list made of 4 elements: minimum longitude (lonMin), minimum latitude (latMin), maximum longitude (lonMax), maximum latitude (latMax)}

\item{stationID}{Station ID number, it should be in the range [1104150,6990700]}

\item{columnName}{name of the column to filter}

\item{columnValue}{value to look for in the column named columnName}

\item{mdDescription}{boolean value. Default is FALSE (no description is printed)}

\item{cached}{boolean value. Default is TRUE (use cached datasets).}
}
\value{
This function returns a data frame made of 8966 rows (gauging stations, as per October 2016) and 44 columns containing metadata.
}
\description{
This function interfaces the Global Runoff Data Centre database which provides river discharge data for about 9000 sites over 157 countries.
}
\examples{
\dontrun{
  # Retrieve the whole catalogue
  x <- catalogueGRDC()

  # Define a bounding box
  bbox <- list(lonMin = -3.82, latMin = 52.41,
               lonMax = -3.63, latMax = 52.52)
  # Filter the catalogue based on bounding box
  x <- catalogueGRDC(bbox = bbox)

  # Get only catchments with area above 5000 Km2
  x <- catalogueGRDC(columnName = "area", columnValue = ">= 5000")

  # Get only catchments within river Thames
  x <- catalogueGRDC(columnName = "river", columnValue = "Thames")
}

}
\author{
Claudia Vitolo
}

