% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOPEX.R
\name{catalogueMOPEX}
\alias{catalogueMOPEX}
\title{Interface for the MOPEX database catalogue}
\usage{
catalogueMOPEX()
}
\value{
This function returns a data frame made of 5 columns: "id" (hydrometric reference number), "name", "location", "Latitude" and "Longitude".
}
\description{
This function interfaces the MOPEX database catalogue (available from ftp://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/) containing 438 daily datasets.
}
\examples{
# Retrieve the MOPEX catalogue
# catalogueMOPEX()

}
\author{
Claudia Vitolo
}

