% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data60UKCatalogue.R
\name{Data60UKCatalogue}
\alias{Data60UKCatalogue}
\title{Interface for the Data60UK database catalogue}
\usage{
Data60UKCatalogue(bbox = NULL, metadataColumn = NULL, entryValue = NULL,
  verbose = FALSE)
}
\arguments{
\item{bbox}{bounding box, a list made of 4 elements: minimum longitude (lonMin), minimum latitude (latMin), maximum longitude (lonMax), maximum latitude (latMax)}

\item{metadataColumn}{name of the column to filter}

\item{entryValue}{value to look for in the column named metadataColumn}

\item{verbose}{if TRUE it returns whether the data is coming from live or cached data sources}
}
\value{
This function returns a data frame made of 5 columns: "id" (hydrometric reference number), "name", "location", "Latitude", "Longitude" and "area".
}
\description{
This function interfaces the Data60UK database catalogue (available from http://www.nwl.ac.uk/ih/nrfa/pub/index.html) containing 61 datasets. Dataset catalogue is available from http://www.nwl.ac.uk/ih/nrfa/pub/data.html.
}
\examples{
# Retrieve the whole catalogue
# Data60UKCatalogue()

# Define a bounding box
# bbox <- list(lonMin=-3.82,latMin=52.41,lonMax=-3.63,latMax=52.52)

# Filter the catalogue
# Data60UKCatalogue(bbox)
# Data60UKCatalogue(metadataColumn="id",entryValue="62001")
}
\author{
Claudia Vitolo
}

