% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Data60UKCatalogue}
\alias{Data60UKCatalogue}
\title{Interface for the Data60UK database catalogue}
\usage{
Data60UKCatalogue(lonMin = -180, lonMax = +180, latMin = -90,
  latMax = +90)
}
\arguments{
\item{lonMin}{Minimum latitude of bounding box}

\item{lonMax}{Maximum latitude of bounding box}

\item{latMin}{Minimum longitude of bounding box}

\item{latMax}{Maximum longitude of bounding box}
}
\value{
This function returns a data frame made of 5 columns: "id" (hydrometric reference number), "name", "location", "Latitude" and "Longitude".
}
\description{
This function interfaces the Data60UK database catalogue (available from http://www.nwl.ac.uk/ih/nrfa/pub/index.html) containing 61 datasets. Dataset catalogue is available from http://www.nwl.ac.uk/ih/nrfa/pub/data.html.
}
\examples{
# Data60UKCatalogue()
# Data60UKCatalogue(lonMin=-3.82,lonMax=-3.63,latMin=52.41,latMax=52.52)
}
\author{
Claudia Vitolo
}

