% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtools-package.R
\docType{data}
\name{Data60UKcatalogue}
\alias{Data60UKcatalogue}
\title{Data set: The Data60UK Catalogue}
\format{
A data frame with 61 stations (rows) and 6 metadata fields (columns).
\describe{
  \item{\code{stationID}}{Station id number.}
  \item{\code{River}}{String describing the river's name.}
  \item{\code{Location}}{String describing the location.}
  \item{\code{gridReference}}{British National Grid Reference.}
  \item{\code{Latitude}}{}
  \item{\code{Longitude}}{}
}
}
\source{
\url{http://nrfaapps.ceh.ac.uk/datauk60/data.html}
}
\usage{
data("Data60UKcatalogue")
}
\description{
The Data60UK catalogue
}
\keyword{datasets}
