% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Data60UKDailyTS}
\alias{Data60UKDailyTS}
\title{Interface for the Data60UK database of Daily Time Series}
\usage{
Data60UKDailyTS(hydroRefNumber, plotOption = FALSE, timeExtent = NULL)
}
\arguments{
\item{hydroRefNumber}{hydrometric reference number}

\item{plotOption}{boolean to define whether to plot the results. By default this is set to TRUE.}

\item{timeExtent}{is a vector of dates and times for which the data should be retrieved}
}
\value{
The function returns a data frame containing 2 time series (as zoo objects): "P" (precipitation) and "Q" (discharge).
}
\description{
This function extract the dataset containing daily rainfall and streamflow discharge at one of the Data60UK locations.
}
\examples{
# Data60UKDailyTS(39015)
}
\author{
Claudia Vitolo
}

