% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Data60UKDailyTS}
\alias{Data60UKDailyTS}
\title{Interface for the Data60UK database of Daily Time Series}
\usage{
Data60UKDailyTS(hydroRefNumber)
}
\arguments{
\item{hydroRefNumber}{hydrometric reference number}
}
\value{
The function returns a data frame containing 2 time series (as zoo objects): "P" (precipitation) and "Q" (discharge).
}
\description{
This function extract the dataset containing daily rainfall and streamflow discharge at one of the Data60UK locations.
}
\examples{
# Data60UKDailyTS(62001)
}
\author{
Claudia Vitolo
}

