% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContent}
\alias{getContent}
\title{Extracts links from web page}
\usage{
getContent(dirs)
}
\arguments{
\item{dirs}{is the url from which links should be extracted}
}
\value{
vector containing all the links in the page
}
\description{
This function extracts all the links in a web page
}
\examples{
\dontrun{
  # Retrieve mopex daily catalogue
  url <- "https://hydrology.nws.noaa.gov/gcip/mopex/US_Data/Us_438_Daily/"
  getContent(dirs = url)
}

}
\author{
Claudia Vitolo
}
