% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grdcCatalogue}
\alias{grdcCatalogue}
\title{Interface for the Global Runoff Data Centre database catalogue}
\usage{
grdcCatalogue(BBlonMin = -180, BBlonMax = +180, BBlatMin = -90,
  BBlatMax = +90, liveData = FALSE, mdDescription = FALSE)
}
\arguments{
\item{BBlonMin}{Minimum latitude of bounding box}

\item{BBlonMax}{Maximum latitude of bounding box}

\item{BBlatMin}{Minimum longitude of bounding box}

\item{BBlatMax}{Maximum longitude of bounding box}

\item{liveData}{bolean value that allows to switch between the online dataset (liveData=TRUE) and the cached one (liveData = FALSE). The online database retrieval might be slower but more up-to-date compared with the cached one (updated twice a year).}

\item{mdDescription}{boolean value. Default is FALSE (no description is printed)}
}
\value{
list of stations within the bounding box
}
\description{
This function interfaces the Global Runoff Data Centre database which provides river discharge data for about 9000 sites over 157 countries.
}
\examples{
# grdcCatalogue(BBlonMin=-3.82,BBlonMax=-3.63,BBlatMin=52.41,BBlatMax=52.52)
}
\author{
Claudia Vitolo
}

