% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtools-package.R
\docType{data}
\name{GRDCcatalogue}
\alias{GRDCcatalogue}
\title{Data set: The GRDC Catalogue}
\format{
A data frame with 9252 stations (rows) and 46 metadata fields.
\describe{
  \item{\code{grdc_no}}{: GRDC station number}
  \item{\code{wmo_reg}}{: WMO region}
  \item{\code{sub_reg}}{: WMO subregion}
  \item{\code{nat_id}}{: national station ID}
  \item{\code{river}}{: river name}
  \item{\code{station}}{: station name}
  \item{\code{country}}{: country code (ISO 3166)}
  \item{\code{lat}}{: latitude, decimal degree}
  \item{\code{long}}{: longitude, decimal degree}
  \item{\code{area}}{: catchment size, km2}
  \item{\code{altitude}}{: height of gauge zero, m above sea level}
  \item{\code{ds_stat_no}}{: GRDC station number of next downstream GRDC station}
  \item{\code{d_start_day}}{: daily data available from day}
  \item{\code{d_start_month}}{: daily data available from month}
  \item{\code{d_start}}{: daily data available from year}
  \item{\code{d_end_day}}{: daily data available until day}
  \item{\code{d_end_month}}{: daily data available until month}
  \item{\code{d_end}}{: daily data available until year}
  \item{\code{d_yrs}}{: length of time series, daily data}
  \item{\code{d_miss}}{: percentage of missing values (daily data)}
  \item{\code{m_start_day}}{: monthly data available from day}
  \item{\code{m_start_month}}{: monthly data available from month}
  \item{\code{m_start}}{: monthly data available from year}
  \item{\code{m_end_day}}{: monthly data available until day}
  \item{\code{m_end_month}}{: monthly data available until month}
  \item{\code{m_end}}{: monthly data available until year}
  \item{\code{m_yrs}}{: length of time series, monthly data}
  \item{\code{m_miss}}{: percentage of missing values (monthly data)}
  \item{\code{t_start}}{: earliest data available}
  \item{\code{t_end}}{: latest data available}
  \item{\code{t_yrs}}{: maximum length of time series, daily and monthly data}
  \item{\code{lta_discharge}}{: mean annual streamflow, m3/s}
  \item{\code{r_volume_yr}}{: mean annual volume, km3}
  \item{\code{r_height_yr}}{: mean annual runoff depth, mm}
  \item{\code{l_im_yr}}{: }
  \item{\code{provider_id}}{: }
}
}
\source{
\url{http://www.bafg.de/GRDC/EN/02_srvcs/21_tmsrs/211_ctlgs/catalogues_node.html}
}
\usage{
data("GRDCcatalogue")
}
\description{
The GRDC catalogue
}
\keyword{datasets}
