% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kgclimateclass}
\alias{kgclimateclass}
\title{Function to identify the updated Koppen-Greiger climate zone (on a 0.1 x 0.1 degrees resolution map).}
\usage{
kgclimateclass(BBlonMin, BBlonMax, BBlatMin, BBlatMax, liveData = FALSE,
  updatedBy = "Peel")
}
\arguments{
\item{BBlonMin}{Minimum latitude of bounding box}

\item{BBlonMax}{Maximum latitude of bounding box}

\item{BBlatMin}{Minimum longitude of bounding box}

\item{BBlatMax}{Maximum longitude of bounding box}

\item{liveData}{bolean value that allows to switch between using online dataset (liveData=TRUE, the online datset might be updated) and offline dataset (liveData = FALSE, this might not be the latest version of the dataset).}

\item{updatedBy}{this can either be Kottek (http://koeppen-geiger.vu-wien.ac.at/) or Peel (http://people.eng.unimelb.edu.au/mpeel/koppen.html)}
}
\value{
List of overlapping climate zones.
}
\description{
Given a bounding box, the function identifies the overlapping climate zones.
}
\examples{
# kgclimateclass(BBlonMin=-3.82,BBlonMax=-3.63,BBlatMin=52.41,BBlatMax=52.52,updatedBy="Peel",liveData=TRUE)
# kgclimateclass(BBlonMin=-3.82,BBlonMax=-3.63,BBlatMin=52.41,BBlatMax=52.52,updatedBy="Peel",liveData=FALSE)
# kgclimateclass(BBlonMin=-3.82,BBlonMax=-3.63,BBlatMin=52.41,BBlatMax=52.52,updatedBy="Kottek",liveData=TRUE)
# kgclimateclass(BBlonMin=-3.82,BBlonMax=-3.63,BBlatMin=52.41,BBlatMax=52.52,updatedBy="Kottek",liveData=FALSE)
}
\author{
Claudia Vitolo
}

