% Generated by roxygen2 (4.0.2): do not edit by hand
\name{KGClimateClass}
\alias{KGClimateClass}
\title{Function to identify the updated Koppen-Greiger climate zone (on a 0.1 x 0.1 degrees resolution map).}
\usage{
KGClimateClass(bbox, updatedBy = "Peel", verbose = FALSE)
}
\arguments{
\item{bbox}{bounding box, a list made of 4 elements: minimum longitude (lonMin), minimum latitude (latMin), maximum longitude (lonMax), maximum latitude (latMax)}

\item{updatedBy}{this can either be Kottek}

\item{verbose}{if TRUE more info are printed on the screen}
}
\value{
List of overlapping climate zones.
}
\description{
Given a bounding box, the function identifies the overlapping climate zones.
}
\examples{
# Define a bounding box
# bbox <- list(lonMin=-3.82,latMin=52.41,lonMax=-3.63,latMax=52.52)

# Get climate classes
# KGClimateClass(bbox)
}
\author{
Claudia Vitolo
}
\references{
Kottek et al. (2006): http://koeppen-geiger.vu-wien.ac.at/. Peel et al. (2007): http://people.eng.unimelb.edu.au/mpeel/koppen.html.
}

