% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtools-package.R
\docType{data}
\name{MOPEXcatalogue}
\alias{MOPEXcatalogue}
\title{Data set: The MOPEX Catalogue}
\format{A data frame with 431 stations (rows) and 12 metadata fields (columns).
\describe{
  \item{\code{id}}{id number.}
  \item{\code{Longitude}}{}
  \item{\code{Latitude}}{}
  \item{\code{V4}}{}
  \item{\code{V5}}{}
  \item{\code{start}}{}
  \item{\code{stop}}{}
  \item{\code{V8}}{}
  \item{\code{V9}}{}
  \item{\code{STATEcode}}{}
  \item{\code{V11}}{}
  \item{\code{name}}{String describing the river's name.}
}}
\source{
\url{ftp://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/}
}
\usage{
data("MOPEXcatalogue")
}
\description{
The MOPEX catalogue
}
\keyword{datasets}

