% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtools-package.R
\docType{data}
\name{MOPEXcatalogue}
\alias{MOPEXcatalogue}
\title{Data set: The MOPEX Catalogue}
\format{A data frame with 431 stations (rows) and 12 metadata fields (columns).
\describe{
  \item{\code{stationID}}{Station id number.}
  \item{\code{longitude}}{}
  \item{\code{latitude}}{}
  \item{\code{elevation}}{}
  \item{\code{V5}}{}
  \item{\code{datestart}}{}
  \item{\code{dateend}}{}
  \item{\code{V8}}{}
  \item{\code{V9}}{}
  \item{\code{state}}{}
  \item{\code{V11}}{}
  \item{\code{basin}}{String describing the river's name.}
}}
\source{
\url{ftp://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/}
}
\usage{
data("MOPEXcatalogue")
}
\description{
The MOPEX catalogue
}
\keyword{datasets}

