% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtools-package.R
\docType{data}
\name{SEPAcatalogue}
\alias{SEPAcatalogue}
\title{Data set: The SEPA Catalogue}
\format{A data frame with 830 observations on the following 8 variables.
\describe{
  \item{\code{idNRFA}}{National River Flow Archive id number.}
  \item{\code{aspxpage}}{Environment Agency gauges id.}
  \item{\code{stationId}}{SEPA station id.}
  \item{\code{River}}{String describing the river's name.}
  \item{\code{Location}}{String describing the location.}
  \item{\code{GridRef}}{British National Grid Reference.}
  \item{\code{Operator}}{The operator's name.}
  \item{\code{CatchmentAreaKm2}}{Area of the catchment.}
}}
\source{
\url{http://pennine.ddns.me.uk/riverlevels/ConciseList.html}
}
\usage{
data("SEPAcatalogue")
}
\description{
The SEPA catalogue
}
\keyword{datasets}

