% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtools-package.R
\docType{data}
\name{SEPAcatalogue}
\alias{SEPAcatalogue}
\title{Data set: The SEPA Catalogue}
\format{
A data frame with 830 observations on the following 8 variables.
\itemize{
  \item{\code{SEPA_HYDROLOGY_OFFICE}}
  \item{\code{STATION_NAME}}
  \item{\code{LOCATION_CODE}}
  \item{\code{NATIONAL_GRID_REFERENCE}}
  \item{\code{CATCHMENT_NAME}}
  \item{\code{RIVER_NAME}}
  \item{\code{GAUGE_DATUM}}
  \item{\code{CATCHMENT_AREA}} in Km2
  \item{\code{START_DATE}}
  \item{\code{END_DATE}}
  \item{\code{SYSTEM_ID}}
  \item{\code{LOWEST_VALUE}}
  \item{\code{LOW}}
  \item{\code{MAX_VALUE}}
  \item{\code{HIGH}}
  \item{\code{MAX_DISPLAY}}
  \item{\code{MEAN}}
  \item{\code{UNITS}}
  \item{\code{WEB_MESSAGE}}
  \item{\code{NRFA_LINK}}
}
}
\source{
\url{http://apps.sepa.org.uk/database/riverlevels/SEPA_River_Levels_Web.csv}
}
\usage{
data("SEPAcatalogue")
}
\description{
The SEPA catalogue
}
\keyword{datasets}
