% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data60UK.R
\name{tsData60UK}
\alias{tsData60UK}
\title{Interface for the Data60UK database of Daily Time Series}
\usage{
tsData60UK(id)
}
\arguments{
\item{id}{String which identifies the station ID number}
}
\value{
The function returns a data frame containing 2 time series (as zoo objects): "P" (precipitation) and "Q" (discharge).
}
\description{
This function extract the dataset containing daily rainfall and streamflow discharge at one of the Data60UK locations.
}
\examples{
\dontrun{
  Morwick <- tsData60UK(id = "22001")
}

}
\author{
Claudia Vitolo
}
