% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEPA.R
\name{tsSEPA}
\alias{tsSEPA}
\title{Interface for the MOPEX database of Daily Time Series}
\usage{
tsSEPA(id)
}
\arguments{
\item{id}{hydrometric reference number (string)}
}
\value{
The function returns river level data in metres, as a zoo object.
}
\description{
This function extract the dataset containing daily rainfall and
streamflow discharge at one of the MOPEX locations.
}
\examples{
\dontrun{
  sampleTS <- tsSEPA(id = "10048")
}

}
\author{
Claudia Vitolo
}
